check of symlinks and directories existence separately

+++ scripts/init.bash
@@ -5,10 +5,8 @@ cd ../../
 ( cd latex-template && vagrant up --provision )
 
 for dir in dia dot gpline inkscape png tex pseudocode umlet; do
-	if [ ! -d "$dir" ]; then
-		mkdir "$dir"
-		ln -s "../../${dir}" "latex-template/src/${dir}"
-	fi
+	[ ! -d "$dir" ]                    && mkdir "$dir"
+	[ -L "latex-template/src/${dir}" ] && ln -s "../../${dir}" "latex-template/src/${dir}"
 done
 
 cat <<EOF > Makefile
